requirements.txt
asgiref==3.2.7
Django==3.0.5
pytz==2019.3
sqlparse==0.3.1
ujson==2.0.3
easy-thumbnails>=2.7
django-image-cropping==1.3.0
Fabric3>=1.14
Pillow>=7.0.*
django-constance==2.6.*
django-cors-headers==3.2.*
djangorestframework==3.11.*
django-rest-auth==0.9.*
django-allauth==0.41.*
